Images with several different GUI options, with dockerfiles split out for easier caching when doing builds
The containers run Xvfb as the X server and expose it via VNC on port
5901; the `launch` script runs the container with that port exposed.

The current `build*` and `launch` scripts build and run with `sudo
podman`; will change to `docker` later.

The xfce-based image is the most developed; runs as an unpriv user
(though has sudo for now so I can experiment).